requirements.txt
nltk==3.3
requests==2.18.4
numpy==1.13.0
gensim==3.6.0
pandas==0.23.4
PyYAML==3.13
